Day 4 - moved Footer.jsx to components/layout
