+++ frontend/src/pages/MyEvents.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import axios from "axios";
 import Navbar from "../components/layout/Navbar";
 import EventCard from "../components/event/EventCard";
-import Footer from "./Footer";
+import Footer from "../components/layout/Footer";
 
 export default function MyEvents() {
   const [events, setEvents] = useState([]);
