+++ frontend/src/pages/ProfilePage.jsx
@@ -3,7 +3,7 @@ import { useNavigate } from "react-router-dom";
 import { updateProfile } from "../api/auth";
 import { getMyOrganization } from "../api/organization";
 import Navbar from "../components/layout/Navbar";
-import Footer from "./Footer";
+import Footer from "../components/layout/Footer";
 
 export default function ProfilePage() {
   const [user, setUser] = useState(null);
