+++ frontend/src/pages/YourOrganizations.jsx
@@ -4,7 +4,7 @@ import React, { useEffect, useState } from "react";
 import axios from "axios";
 import { useNavigate } from "react-router-dom";
 import Navbar from "../components/layout/Navbar";
-import Footer from "./Footer";
+import Footer from "../components/layout/Footer";
 
 export default function YourOrganizations() {
   const [orgs, setOrgs] = useState([]);
